+++ Codigo mastermind/GameRules.java
