--- Mastermind/nbproject/genfiles.properties
@@ -1,8 +1,0 @@
-build.xml.data.CRC32=7ee6d75b
-build.xml.script.CRC32=269a8bfb
-build.xml.stylesheet.CRC32=8064a381@1.75.1.48
-# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
-# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
-nbproject/build-impl.xml.data.CRC32=7ee6d75b
-nbproject/build-impl.xml.script.CRC32=98ec45a0
-nbproject/build-impl.xml.stylesheet.CRC32=876e7a8f@1.75.1.48
