--- Mastermind/nbproject/private/private.properties
@@ -1,2 +1,0 @@
-compile.on.save=true
-user.properties.file=C:\\Users\\Gabri\\AppData\\Roaming\\NetBeans\\8.0.1\\build.properties
